Introduce CreationsList and CreationsListEntry
The CreationsListView is now a view on CreationsList, and CreationsList
is a list of CreationsListEntry's, and a CreationsListEntry is a list of
requested name-words along with the Creations they are associated with.

CreationsListView <>--> CreationsList <>--> CreationsListEntry <>--> Creation <>--> Recording

Left to do: selection should be based on entries, not recordings. That
will remove the need of dummy recordings, and CreationsListEntry's will
represent what the user wants, which may not necessarily be available
yet (e.g. concatenated names).