Relax selection logic to include CreationListEntries
Previously, the select-next and select-previous was broken when multiple
non-concatenated names were selected, as the selection does not exactly
match any single recording.

Now, instead of trying to match the CreationListEntries to the selection
exactly, we relax the selection condition such that the
CreationListEntry is selected when it is found as a sub-sequence in the
selection.

The checkboxes now work as you'd expect: selecting a MultiCellContents
via the checkbox does not deselect the current selection.